Refactor: Consolidate code for query generator
