--- public/app/features/trails/AutomaticMetricQueries/query-generators/common/index.ts
@@ -1,9 +1,0 @@
-import { generateQueries } from './queries';
-import { getGeneratorParameters } from './rules';
-
-function generator(metricParts: string[]) {
-  const params = getGeneratorParameters(metricParts);
-  return generateQueries(params);
-}
-
-export default { generator };
