--- public/app/features/trails/AutomaticMetricQueries/query-generators/common/types.ts
@@ -1,5 +1,0 @@
-export type AutoQueryParameters = {
-  agg: string;
-  unit: string;
-  rate: boolean;
-};
