--- public/app/features/trails/AutomaticMetricQueries/query-generators/types.ts
@@ -1,3 +1,0 @@
-import { AutoQueryInfo } from '../types';
-
-export type MetricQueriesGenerator = (metricParts: string[]) => AutoQueryInfo;
